fixed ascii reading

+++ NeuralNetwork/headers/NeuralNetworkTools.h
@@ -32,7 +32,7 @@ struct NeuralNetwork
 double randd();
 
 void freeNetwork(NeuralNetwork* network);
-NeuralNetwork* GenerateNetwork(size_t inputNumber, size_t hidenNumber, size_t outputNumber, int * asciiOutputs);
+NeuralNetwork* GenerateNetwork(size_t inputNumber, size_t hidenNumber, size_t outputNumber, double * asciiOutputs);
 void printList(double* list, size_t len);
 void printNetwork(NeuralNetwork* network);
